Bug 512398 - Backed out changeset f4087c5ac148, checked in wrong patch.
UltraBlame original commit: 8c409bffb344cda4a8118812861c4f72c58b8e01

+++ media/liboggplay/src/liboggplay/oggplay_callback.h
@@ -57,7 +57,7 @@ oggplay_process_leftover_packet(OggPlay *me);
 
 
 OggPlayDecode *
-oggplay_initialise_decoder(OggPlay *me, int content_type, long serialno);
+oggplay_initialise_decoder(OggPlay *me, int content_type, int serialno);
 
 int
 oggplay_callback_info_prepare(OggPlay *me, OggPlayCallbackInfo ***info);
